[FEATURE, FIX] Finally the browser does offer full copy/paste support.
Normal paste actions are treated by the default paste handler. File and image paste actions are delegated to a custom paste handler that circumvents the problem that files get inserted by a string representing their paths. Images in Safari are pasted with a wekit-fake-url - a URL that is only valid in the running browser session.
The custom handler reads the corresponding images and includes a data-url based image.