[FIX,FEATURE] ItemList click events could return the wrong element thus no action is activated.
Solved with the following feature. The browser mouse events provide the element the event originally occurred on.
This can be an inner tag element although it was expected that the outer element was passed.
You can provide the right element by adding one of the following classes to make the browser select the parent element: no_mouseup, no_mousedown, no_click.
Example:
<a><b class="no_click"><c class="no_click"></c></b></a>
If a user clicks on element c, a is part of the event.